requirements.txt
ipdb
ipykernel
yapf
matplotlib
pytest
seaborn
tensorflow
fastapi
uvicorn
tensorflow
pillow
numpy==1.23.5
pandas==1.5.3
scipy==1.10.0
scikit-learn==1.3.1
google-cloud-bigquery
google-cloud-storage==2.14.0
google-api-core==2.8.2
googleapis-common-protos==1.56.4
protobuf==3.19.6
h5py==3.10.0
db-dtypes
pyarrow
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
